// src/app/types/work.ts
export interface Work {
    id: string;
    title: string;
    imageUrl: string;
    imageUrl1: string;
    imageUrl2: string;
    imageUrl3: string;
    imageUrl4: string;
    imageUrl5: string;
    imageUrl6: string;
    imageUrl7: string;
    imageUrl8: string;
    description: string;
  }
  

  